[vim] remove frecency, [home] remove sqlite
Never got to use frecency properly. It was more in the way rather than a
part of the workflow.

The main issue is that files should be prioritized differently:
- Frecency INSIDE the project
- Other files in the project (the ones that were never edited)
- Not sure if it should display files outside of project (or CWD)
- Pretty much the vscode behavior

Sqlite was needed in home just for frecency, so removing it to unclutter
and keep the system clean. Projects that need sqlite can easily include
it in shell.nix.