[vim][home.nix] w000t bootstrapped vim!
This nukes the use of vim-packer and utilizes nix to manage plugins.
The problem I was running into was the initial chicken-egg error swamp
on a fresh install. Vim would load vimrc, but no package manager was
installed, thus you couldn't install packages.

The workflow was:
  1. Install OS, or do a major update
  2. Launch Vim, swim in errors
  3. Search the interwebs how to install vim-packer
  4. Git clone vim-packer somewhere
  5. Fire up vim again, swim in errors again.
  6. :PackerSync
  7. Profit

Now the workflow is:
  1. Install/update OS. nixos-rebuild --switch
  2. There's no step 2.